+++ src/components/Layout/index.tsx
@@ -5,10 +5,19 @@ import MeetGoalIcon from "../../icons/MeetGoalIcon";
 import { Navigation } from "../Navigation";
 import MenuIcon from "../../icons/MenuIcon";
 import FooterActions from "../Footer/index";
+import { useLocation } from "react-router-dom";
+import { animateScroll as scroll } from "react-scroll";
+import { Routes } from "../../constants/routes";
+import { CaretUpOutlined } from "@ant-design/icons";
 
 const { Header, Content, Footer } = Layout;
 
 export const MainLayout: React.FC = ({ children }) => {
+  let location = useLocation();
+
+  function goToTop() {
+    scroll.scrollToTop();
+  }
   return (
     <Layout>
       <Header className="header">
@@ -34,6 +43,28 @@ export const MainLayout: React.FC = ({ children }) => {
       <Footer className="footer">
         <FooterActions />
       </Footer>
+      {location.pathname === Routes.HOME ||
+      location.pathname === Routes.SERVICES ||
+      location.pathname === Routes.TEAM ? (
+        <Button
+          style={{
+            position: "fixed",
+            bottom: "5%",
+            right: "5%",
+            width: "50px",
+            height: "50px",
+            borderColor: "#99ff98",
+            background: "transparent",
+          }}
+          icon={
+            <CaretUpOutlined style={{ fontSize: "40px", color: "#99ff98" }} />
+          }
+          shape="circle"
+          onClick={() => goToTop()}
+        ></Button>
+      ) : (
+        ""
+      )}
     </Layout>
   );
 };
